Move shouldPerformDiffInformedAnalysis()
